Kill another signed/unsigned warning.
git-svn-id: http://svn.boost.org/svn/boost/trunk@24005 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/mpl/test/for_each.cpp
@@ -80,7 +80,7 @@ int main()
     mpl::for_each< numbers >(value_printer(std::cout));
     
     for (unsigned i = 0; i < v.size(); ++i)
-        assert(v[i] == i);
+        assert(v[i] == (int)i);
 
     return 0;
 }
